requirements.txt
# Core Dependencies
pandas>=1.5.0,<3.0.0
numpy>=1.21.0,<2.0.0
matplotlib>=3.5.0,<4.0.0
seaborn>=0.11.0,<1.0.0

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=0.19.0
pathlib2>=2.3.0
typing-extensions>=4.0.0

# CLI and User Interface
click>=8.0.0
rich>=12.0.0
tqdm>=4.64.0

# Data Validation and Quality
cerberus>=1.3.4
pandas-profiling>=3.6.0
great-expectations>=0.15.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.2.0

# Code Quality
black>=22.0.0
flake8>=5.0.0
isort>=5.10.0
mypy>=1.0.0
pre-commit>=2.20.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.2.0
myst-parser>=0.18.0

# Development Tools
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.0.0

# Export and Reporting
openpyxl>=3.0.0
xlsxwriter>=3.0.0
reportlab>=3.6.0
jinja2>=3.1.0

# Performance and Monitoring
psutil>=5.9.0
memory-profiler>=0.60.0

# Optional: Database Support
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0

# Optional: Cloud Storage
boto3>=1.26.0
google-cloud-storage>=2.7.0